requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.1
django-celery-beat==2.5.0
celery==5.3.0
redis==4.5.5
Pillow==10.0.0
python-decouple==3.8
PyYAML==6.0
drf-yasg==1.21.4